fix merged code
