+++ blockchain/config.go
@@ -32,6 +32,9 @@ type (
 		BloomfilterIndexDBPath string           `yaml:"bloomfilterIndexDBPath"`
 		CandidateIndexDBPath   string           `yaml:"candidateIndexDBPath"`
 		StakingIndexDBPath     string           `yaml:"stakingIndexDBPath"`
+		SGDIndexDBPath         string           `yaml:"sgdIndexDBPath"`
+		SGDIndexCacheSize      int              `yaml:"sgdIndexCacheSize"`
+		SGDContract            string           `yaml:"sgdContract"`
 		ID                     uint32           `yaml:"id"`
 		EVMNetworkID           uint32           `yaml:"evmNetworkID"`
 		Address                string           `yaml:"address"`
@@ -83,6 +86,8 @@ var (
 		BloomfilterIndexDBPath: "/var/data/bloomfilter.index.db",
 		CandidateIndexDBPath:   "/var/data/candidate.index.db",
 		StakingIndexDBPath:     "/var/data/staking.index.db",
+		SGDIndexDBPath:         "/var/data/sgd.index.db",
+		SGDIndexCacheSize:      1000,
 		ID:                     1,
 		EVMNetworkID:           4689,
 		Address:                "",
